@@ -1,4 +1,4 @@
-class CreateTaxonomies < ActiveRecord::Migration
+class CreateTaxonomies < ActiveRecord::Migration[4.2]
   def change
     create_table :families do |t|
       t.string :name, null: false, unique: true
